+++ src/pages/Construct/useConstructPage.ts
@@ -1,18 +1,22 @@
-import { useSelector } from 'react-redux';
-
 import { selectHasCardData } from '@/redux/cards/cardsSelectors';
-import { selectHasDeckData } from '@/redux/deck/deckSelectors';
-import { selectHasPoolData } from '@/redux/pool/poolSelectors';
+import { useAppSelector } from '@/redux/hooks';
+import {
+  selectAllCardsInPool,
+  selectHasPoolData,
+} from '@/redux/pool/poolSelectors';
 import { selectHasSetData } from '@/redux/sets/setsSelectors';
 
 const useConstructPage = () => {
-  const hasSetData = useSelector(selectHasSetData);
-  const hasCardData = useSelector(selectHasCardData);
-  const hasPoolData = useSelector(selectHasPoolData);
-  const hasDeckData = useSelector(selectHasDeckData);
-  const hasAccess = hasSetData && hasCardData && hasPoolData && !hasDeckData;
+  // hasAccess
+  const hasSetData = useAppSelector(selectHasSetData);
+  const hasCardData = useAppSelector(selectHasCardData);
+  const hasPoolData = useAppSelector(selectHasPoolData);
+  const hasAccess = hasSetData && hasCardData && hasPoolData;
 
-  return { hasAccess };
+  // allCardsInPool
+  const allCardsInPool = useAppSelector(selectAllCardsInPool);
+
+  return { hasAccess, allCardsInPool };
 };
 
 export default useConstructPage;